+++ vite.config.ts
@@ -74,7 +74,7 @@ export default defineConfig({
       // specifier and resolves correctly.
       "libsodium-wrappers-sumo": path.resolve(
         __dirname,
-        "node_modules/libsodium-wrappers-sumo/dist/modules-sumo/libsodium-wrappers.js"
+        "node_modules/libsodium-wrappers-sumo/dist/modules-sumo/libsodium-wrappers.js",
       ),
     },
   },